updated explore file, acquire file and prep file

+++ acquire.py
@@ -13,18 +13,22 @@ def get_store_data():
     '''
     Returns a dataframe of all store data in the tsa_item_demand database and saves a local copy as a csv file.
     '''
-    query = '''
-    SELECT *
-    FROM items
-    JOIN sales USING(item_id)
-    JOIN stores USING(store_id) 
-    '''
+    filename = 'tsa_store_data.csv'
+    if os.path.isfile(filename):
+        return pd.read_csv(filename).iloc[:,1:]
+    else:
+        query = '''
+        SELECT *
+        FROM items
+        JOIN sales USING(item_id)
+        JOIN stores USING(store_id) 
+        '''
     
-    df = pd.read_sql(query, get_db_url('tsa_item_demand'))
+        df = pd.read_sql(query, get_db_url('tsa_item_demand'))
     
-    df.to_csv('tsa_store_data.csv', index=False)
+        df.to_csv('tsa_store_data.csv', index=False)
     
-    return df
+        return df
 
 def wrangle_store_data():
     '''
